server/src/requirements.txt
google_api_python_client==1.6.4
google_auth_httplib2==0.0.2
jinja2==2.10.3
oauth2client==4.1.2
oauthlib==2.0.6
python-dateutil==2.8.1
google-cloud-datastore==1.13.2
requests==2.22.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
PyJWT==1.7.1
PyYAML==5.3.1
validate-email==1.3
validators==0.12.0
